Decode QUOTED-PRINTABLE strings (in names)
There are a handful of JS libs that can do this, but they all have a ton
of potentially excess code we'd never use, they're all >= 8 years old
and unmaintained, and at least one of them doesn't even do multibyte.

Maybe reconsider a lib later (for encoding especially?) but for now, as
long as there's no partially-encoded strings, this does the trick.