Show error in console for unsupported VCF lines
I've had a variation of this since the beginning that would add raw data
to a `<pre>` element which was then appended to the contact.

This need for raw data when a weird line shows up will never go away,
and since I accidentally committed the setting of `hasInvalidLines` in
the parsing switch's default statement, we may as well make it concrete.