Parse property parameters when exporting data
If both name *and* value are set then we join them with `=`, but no
assumption is made about the key/value if only one field is defined.

For instance, `{name: 'PREF', value: ''}` could be converted to
'PREF=1', but what if someone wanted a value of 50? Likewise other
values shouldn't be assumed to be `TYPE=...` so new format will match
the old one when one or more contacts are exported to a new VCF file.