Autodetect length of folded lines for export
Lines can be folded at any length so we shouldn't just assume 64 chars.
This commit adds *some* autodetection, but it still assumes that all
lines of a given property are folded at the same detected length.

That might not always be the case, but for now it does the job.