Fix dragging by image, cleanup contact moving code
Dragging the contact image would have browser handling take over to an
extent, ie the drop event would still fire but the intitial target would
be the `<img>` element rather than the `li.contact` it's supposed to be.

Disabling `draggable` on the img element seems to be enough to solve it.