Allow for dropping contacts on vcard headers
This adds detection of the parent VCard column so we don't need to rely
on arbitrary numbers of nested `.parentNode` properties and can open up
the drop target to the entire VCard column, header and all.

As an added bonus, this also fixes—completely by accident—the issue
where a contact couldn't be dropped if it was on top of another contact.

Yay!