Allow for rearranging contacts within a column
Note this is purely visual and the object could be in any order
whatsoever. I wrote a bunch of stuff for recreating the contacts,
praying it'd iterate in the new order, but it didn't. Oh well.

Maybe redo contacts using arrays and have helper methods to find by id
keys on the value objects instead, but for now we can live without.