Fix bug when dropping contacts in empty space
If you dropped a contact in the empty space created above the
last-hovered contact, it wouldn't be able to identify that contact,
instead inserting the new contact at the top of the vCard as if it were
dropped on the vCard's header element rather than above another contact.

This updates it to offset the cursor position by the last-hovered height
and grabs the contact element from that position instead where possible.