Fix vCard column removal
Fixes an issue with the delete button on vCard columns, introduced when
icons were converted to the material design outline font. The extra
level of nesting caused by the `i` meant only the button was removed.

This updates the event handler to accept the vCard ID and use that to
grab the instance from `this.vCards` and remove via its column property.

Doing it this way also enables easy removal of the actual array item, so
we kill two birds with one stone and ensure full removal of the vCard.