Prefer to show phone or email in contact headers
If any phone or email properties exist, the first match will be shown on
the second line of the contact header in place of any company details.

This also adds a further fallback that will display the first property,
no matter what it is, if none of those properties are available.