Don't show email if it's used as primary
Updates the vCard generator to pass the current 'primary' value along to
the 'secondary' method so potential values can be checked against it.

If an email is the first property and also used as the primary info, the
code would run through to the final return and show the email anyway. To
avoid that we have to filter the final return against the primary also.

We also filter the properties using the `OnlyExtraPropertiesFilter` to
avoid the value of basic properties like `BEGIN:VCARD` being used.