Greatly improve contact reordering experience
Previously when an element was being dragged over, a 'hovering' class
was added which would add a bunch of space above the contact as a
dropzone to avoid merging into it when only sorting was desired.

This instead utilises dragenter (mostly to avoid continuous calls) and
compares the top offsets of both the dragged and currently-hovered
contacts to see which is first at that given moment in time.

If the *dragged* contact appears first, it will be inserted *before* the
contact being hovered, and vice versa. This creates a "buffer" allowing
the dragged contact to be dropped on the contact above/below it.

Inverting the greater-than would cause instant sorting. A desirable
feature in some use cases, but it would prevent merging contacts.