Allow drop outside contact but still within vcard
A lot less actual changes than you'd thing, most are a result of storing
both card and contact divs in the `dragging` class property and adding
an assertion function for it because I was sick of "it might be null".

The real meat of this is in the `DragEnter` and `Drop` handlers where if
statements are added/tweaked to allow for dropping a contact in another
vCard column, while not necessarily being on a contact in that column.

This re-enables moving to an empty vCard and also fixes a regression
that broke contact merging if a contact moved to a new card column.