Also reset position on merge
We're not so much wanting to track the *drag* completing, as the *move*.

If we start a merge operation, it doesn't make sense to also complete
the move as that's likely not the intent. Setting `didMove` (which seems
more appropriate than `didDrag`) in the move method solves the trick.

The one potential downside of this approach is it also resets position
when dropped in the empty space around contacts but still within the
vcard column. Though since that doesn't trigger an actual contact move
between the columns' `contacts` arrays, maybe that's a good thing.