Improve moving to cards with few contacts
If a card only had a small number of contacts, positioning during the
drag was a bit... sporadic. Likewise the contact would always be
appended, regardless of where it was dropped (if not on a contact).

This adds a function that compares distance from cursor to the centre of
each `.contact` in the target vcard column, returning the smallest one.
If it matches, we add after that instead of appending to `.contacts`.