Allow for influencing insert position above/below
After wasting an entire day yesterday trying to fix too many things at
once, this takes another angle and finally fixes issues with dragging
where you'd be unable to rearrange the top items or move an item to the
top of a different contact card column.

The drag enter event is left in tact for non-contact (ie vcard) events,
whereas the drag*over* event is used to detect whether the cursor is
above or below the halfway point of the hovered' contact's height.