Skip nearby detection for contacts outside view
If the bottom of the contact is higher than the top of the `<ul>` (that
is, the visible part of the contacts list excluding header) then we can
skip distance calculations to save a little time.

If the *top* of the contact is lower than... you get the idea.. then
rather than skip it, we can break out of the loop entirely as we know
it'll always be looping through in order.

Originally had it returning early from the `forEach()`, but `return` is
ALL it can do, so switching it out for a real `for` loop seems better.

Could probably avoid the bottom check by instead checing if distance is
suddenly *further* than the shortest distance (which can only be true
once it's had a value set that's not `Infinity`), but this'll do.

Likewise I could also improve the top part by splitting it in half in
case we're scrolled way down near the end of the list, but there's no
sense in that as long as it's only dealing with 100 or so contacts.