Clean up observers, ensure they remain after moves
If a contact is moved to a different VCard, it would be detected as no
longer visible in the old card and—if later restored—be included again.

However, it would never be detectable in the new card if we don't also
observe the contact there. Since this wraps contact add/remove for a
vCard instance, we also wrap the delete to have observers disconnect.
Probably isn't necessary, but it saves relying on garbage collection.