Backtrack a bit
Using the Entry's bounding rect was a wonderful idea, 'til it wasn't.

In hindsight it's obvious that it would be outdated by the time a
contact is dragged about, so we end up with odd behaviour where most of
the contacts will all be thought to be super close at the top/bottom,
and we end up with a disco if you slowly drag up/down over the header.

This reverts pretty much all of that, short of a variable rename and
passing the entry's target directly instead of calling `ui.element`.