Add simple phone number formatting
All it does is remove anything that's not a digit, plus, hash or star.

No accounting for where the '+' is, and it doesn't care for brackets or
dots/commas used to visually separate the area code etc either. Could
eventually be made to be far more flexible, but this solves my needs.